requirements.txt
gradio>=4.31.5
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
google-cloud-documentai>=2.20.0
google-cloud-storage>=2.14.0
pandas>=2.1.0
numpy>=1.26.0
python-docx>=1.1.0
pypdf>=3.17.0
firebase-admin>=6.4.0
SQLAlchemy>=2.0.0
cloud-sql-python-connector>=1.0.0
pg8000>=1.30.0
